requirements.txt
black==24.4.2
Django==4.2.11
djangorestframework==3.15.1
